+++ Changes.md
@@ -4,7 +4,9 @@
 Improvements
 ------------
 
-- OSLShader : Improved loading of spline parameters with additional duplicate endpoints (including some RenderMan shaders).
+- OSLShader :
+ - Improved loading of spline parameters with additional duplicate endpoints.
+ - Added support for loading splines from RenderMan shaders.
 
 Fixes
 -----